Updated Controllers.
Added parameterless constructor (Default parameters isn't enough for
WebApi).
Updated Extensions to use Ids instead of object-references.

+++ code/DriveIT/DriveIT.WebAPI/Controllers/CommentsController.cs
@@ -11,15 +11,19 @@ namespace DriveIT.WebAPI.Controllers
     {
         private readonly IPersistentStorage _repo;
 
-        public CommentsController(IPersistentStorage repo = null)
+        public CommentsController(IPersistentStorage repo)
         {
-            _repo = repo ?? new EntityStorage();
+            _repo = repo;
         }
+
+        public CommentsController() : this(new EntityStorage()) { }
 
         // GET: api/Comments/5
         // Where 5 is CarId
-        public async Task<IHttpActionResult> Get(int carId)
+        public async Task<IHttpActionResult> Get(int id)
         {
+            // Remember: It is a car-id. Not commentId.
+            var carId = id;
             var comments = (from comment in await _repo.GetAllCommentsForCar(carId)
                             select comment.ToDto()).ToList();
             if (!comments.Any())
@@ -36,25 +40,38 @@ namespace DriveIT.WebAPI.Controllers
             {
                 return BadRequest(ModelState);
             }
-            var newCommentId = await _repo.CreateComment(await value.ToEntity(_repo));
+            if (value == null)
+            {
+                return BadRequest("Null value not allowed.");
+            }
+            var newCommentId = await _repo.CreateComment(value.ToEntity());
             return CreatedAtRoute("DefaultApi", new { id = newCommentId }, value);
         }
 
         // PUT: api/Comments/5
         public async Task<IHttpActionResult> Put(int id, [FromBody]CommentDto value)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var comment = await _repo.GetCommentWithId(id);
             if (comment == null)
             {
                 return NotFound();
             }
-            await _repo.UpdateComment(id, await value.ToEntity(_repo));
+            await _repo.UpdateComment(id, value.ToEntity());
             return Ok();
         }
 
         // DELETE: api/Comments/5
         public async Task<IHttpActionResult> Delete(int id)
         {
+            var comment = await _repo.GetCommentWithId(id);
+            if (comment == null)
+            {
+                return NotFound();
+            }
             await _repo.DeleteComment(id);
             return Ok();
         }
